pyproject.toml
[project]
name = "unblind"
version = "0.1.0"
description = "Extract math problem statements from images via Gemini 2.5 Flash."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-genai>=1.0",
    "python-dotenv>=1.0",
]
